expose agent capabilities in global clientCapabilities()
Previously, only the agent code could read the agent capabilities, but they're actually useful throughout the Cody codebase and describe what the current client is capable of. We should use them instead of checking `agentIDE` and assuming based on what kind of editor it is on the other side, since that is coarse.